signing vouchers is done with the key provider from the config, and only applies to vouchers
