+++ examples/ore/sign_vouchers.js
