+++ test/instrument.test.js
@@ -193,4 +193,12 @@ describe('instrument', () => {
       expect(stats).toEqual({ totalCpuUsage: totalCpu * 2, totalApiCalls: totalCount * 2 });
     });
   });
+
+  describe('signVoucher', () => {
+    test('signs a voucher', async () => {
+      const voucherId = 0;
+      const sig = await orejs.signVoucher(voucherId);
+      expect(sig.toString()).toEqual('SIG_K1_K7SnTcWTVuatvRepJ6vmmiHPEh3WWEYiVPB1nD9MZ3LWz91yUxR5fUWmSmNAAP9Dxs2MeKZuDUFoEVfBiKfRozaG2FzfvH');
+    });
+  });
 });
